O 02 - Printing the Linked List in Reverse
You are given the pointer to the head node of a linked list and you need to print all its elements in reverse order from tail to head, on one line separated by space. The head pointer may be null meaning that the list is empty - in that case, do not print anything!

Input Format

You have to complete the void ReversePrint(LinkedListNode* head) method which takes one argument - the head of the linked list. You should NOT read any input from stdin/console.

Output Format

Print the elements of the linked list in reverse order to stdout/console (using printf or cout) , separated by a space.

Lang C