N D04 - Harry Potter and the Prisoner of Azkaban
Harry is riding buckbeak for the first time. While enjoying the blissful ride in the air he notices two numbers represented as integer-digit linked lists flying in the sky (Strange things happen in the wizarding world). Harry is a fan of bigger numbers and hence he wants to add the two flying linked lists.

image
Harry is a wizard and not a programmer and so he does not know how to add two singly linked lists represented as numbers. Complete the task for him.

Formally, write a function that takes two linked lists as input and returns a linked list that represents the sum of both linked lists,