CCT B1 - Harry Potter and the Deathly Hallows
Harry is in his final battle against Lord Voldemort.
Meanwhile Neville Longbottom has found the sword of Godric Gryffindor and plans to kill nagini with it.

Neville is nervous as he is holding the sword for the first time. He notices that nagini looks like a linked list of integers ( .... ).

image
He decides that making you write a code will be simpler than killing the snake with a sword.

Nagini as a snake is a linked list and to kill it Nevil wants you to slice off (delete) exactly one element from it. Neville knows the snakes weakest point x and hence wants you to delete the xth element from the linked list only. He decides to give you the pointer to the element that he wants to delete.

Formally, write a function that deletes the xth element of a linked list given the pointer to it. If deleting the xth node is not possible do not modify the original list.